crypto/requirements.txt
alpaca-py
pandas
numpy
python-dotenv
pytz
flask
